+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added page to edit user profile
 - Added german translation files by Senuros (https://github.com/Senuros)
 - Update to A-Frame 0.8.2
+- Added setting: show lost password link on login form 
 
 # 1.1.4